azure_function/recommendation-function/requirements.txt
azure-functions
azure-storage-blob
numpy==1.26.4
scikit-surprise==1.1.4
pandas
pickle5